requirements.txt
pandas==2.2.3
python-dotenv==1.0.1
requests==2.32.3
jinja2==3.1.4
ttkbootstrap==1.10.1
rich==13.9.4
openpyxl==3.1.5
PyYAML==6.0.2
scikit-learn==1.5.2
joblib==1.4.2
# IA opcional (puedes quitarlo si no lo usarás)
openai==1.51.2
matplotlib==3.9.*
tkcalendar>=1.6
pywin32>=306
numpy>=1.26
